Modify argument.

+++ chapter07/fp.cpp
@@ -2,9 +2,9 @@
 
 int main()
 {
-    double a = -1.5e38, b = 1.5e38, c = 1.0;
-    double d = a + (b + c);
-    double e = a + b + c;
+    float a = -1.5e38, b = 1.5e38, c = 1.0;
+    float d = a + (b + c);
+    float e = a + b + c;
 
     printf("%f\n", d); // 0.000000
     printf("%f\n", e); // 1.000000
